pyproject.toml
[project]
name = "intuitive-care"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "pytest>=8.3.5",
    "requests-mock>=1.12.1",
    "responses>=0.25.7",
    "types-requests>=2.32.0.20250328",
]

